+++ service/src/main/java/bio/terra/externalcreds/controllers/OidcApiController.java
@@ -12,7 +12,6 @@ import bio.terra.externalcreds.services.JwtUtils;
 import bio.terra.externalcreds.services.LinkedAccountService;
 import bio.terra.externalcreds.services.PassportProviderService;
 import bio.terra.externalcreds.services.PassportService;
-import bio.terra.externalcreds.services.TokenProviderService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -32,7 +31,6 @@ public record OidcApiController(
     OauthApiController oauthApiController,
     ObjectMapper mapper,
     PassportService passportService,
-    TokenProviderService tokenProviderService,
     PassportProviderService passportProviderService,
     ExternalCredsSamUserFactory samUserFactory)
     implements OidcApi {
